+++ cli/locale.go
@@ -1,9 +1,9 @@
-package internal
+package cli
 
 import (
 	"reflect"
 
-	"github.com/renatopp/go-cli/locales"
+	"github.com/renatopp/go-cli/cli/locales"
 )
 
 // Locale is re-exported here for convenience of the internal package.