+++ examples/list/list.go
@@ -3,14 +3,14 @@ package main
 import (
 	"fmt"
 
-	"github.com/renatopp/go-cli"
+	"github.com/renatopp/go-cli/cli"
 )
 
 func main() {
 	cli.Name("list")
 	cli.Description("List folders and files")
-	cli.Command("folders", "List folders", cmdFolders)
-	cli.Command("files", "List files", cmdFiles)
+	cli.Cmd("folders", "List folders", cmdFolders)
+	cli.Cmd("files", "List files", cmdFiles)
 
 	// if a command is provided, parse will exit after executing it, so the code
 	// after this won't be executed.